refactor(worker): split cmd/worker/main.go into per-concern files (#410)
main.go had grown to 1645 lines: the process entry/wiring plus the entire
SPEC §13.7 state HTTP server and its JSON API surface in one file. Split it
into three files in the same package main, moving whole declarations
verbatim (no logic change):

  main.go (569)              entry, flag parsing, run() wiring, startup
                             workflow resolution, tracker-client wiring
  statehttp_server.go (438) §13.7 server lifecycle: readiness gate, the
                             hot-reloadable controller + run loop, server
                             construction, loopback/auth guard, probes, and
                             the desiredPort/HostForLoop helpers the loop uses
  stateapi.go (613)         §13.7 API surface: JSON response types, the
                             /api/v1 handlers, StateView -> API mappers, and
                             the refresh-request constants + size-limit
                             sentinel the refresh handler validates against

Kept in package main (additional files), not lifted into a new
internal/worker/bootstrap package: main_test.go (2643 lines) exercises these
unexported symbols, and a same-package move is behavior-risk-free and leaves
the test suite untouched, whereas a new package would force exporting the API
surface and relocating tests — a larger architectural change outside this
file-size refactor.

Declarations are grouped by their consumers: the refresh constants/sentinel
sit with the refresh handler in stateapi.go, and desiredPort/HostForLoop sit
with their sole caller runStateHTTPServerLoop in statehttp_server.go.

Pure structural move. Verified behavior-preserving:
- top-level declaration set identical before/after (101 == 101);
- go test -race ./cmd/worker/ passes;
- funlen/gocognit finding set byte-identical to main (9 == 9, no new
  violations);
- gofmt/go vet/build clean.

Addresses #410 (cmd/worker/main.go file-size hotspot).
